Improve log entry batching
The existing implementation does not really do batching, even though we can set a batch size of 50 that is only ever considered as a maximum, e.g if only 1 log entry has been written we will still process it fully. Given the amount of logging on reasonably used systems this really means we are constantly pushing a small number of logs.

This change respects the batch size fully and does not push logs or update the tracking cursor until a number of logs equal to the batch size has accumulated.